Thumbnailer: add optional input file name regex setting
With the help of the THUMBNAIL_INCLUDE_REGEX setting it is now possible
to restrict thumbnailing to a files matching this regular expression.
The setting is expected to be a string. If available, a regular
expression is created. The default behavior is still that all files not
starting with a dot are respected.

For instance, this can be set to only match files ending in "jpg",
"jpeg" or "png" one can use:

THUMBNAIL_INCLUDE_REGEX = r'(jp[e]g|png)$'